requirements.txt
kafka-python
schedule
faker
psycopg2-binary
